Country metadata from the WHO, UN, and WB.

A dataset containing names, codes, and other attributes for countries included in World Health Organization, United Nations and World Bank datasets.

Usage

countries

Format

A data frame with 249 rows and 109 variables:

iso3

ISO3 code.

iso2

ISO2 code.

iso_numeric

ISO numeric code.

who_code

World Health Organization code.

m49

United Nations M49 code.

gbd_code

IHME GBD country code.

sovereign_iso3

ISO3 code of sovereign nation. Value is NA if sovereign itself.

who_member

Logical, is country a WHO member state?

who_member_small

Logical, is country a WHO member state with population <= 90,000?

oecd_member

Logical, is country an OECD member state?

gbd_high_income

Logical, is the country classified as high-income in the IHME's GBD 2019?

un_ldc

Logical, is the country classified as a least-developed country by the United Nations?

un_lldc

Logical, is the country classified as a land-locked developing country by the United Nations?

un_sids

Logical, is the country classified as a small island developing state by the United Nations?

who_short_name_en

WHO country name, short, English.

who_formal_name_en

WHO country name, formal, English. Only available for official member states.

who_short_name_ar

WHO country name, short, Arabic.

who_formal_name_ar

WHO country name, formal, Arabic. Only available for official member states.

who_short_name_es

WHO country name, short, Spanish.

who_formal_name_es

WHO country name, formal, Spanish. Only available for official member states.

who_short_name_fr

WHO country name, short, French.

who_formal_name_fr

WHO country name, formal, French. Only available for official member states.

who_short_name_ru

WHO country name, short, Russian.

who_formal_name_ru

WHO country name, formal, Russian. Only available for official member states.

who_short_name_zh

WHO country name, short, Chinese.

who_formal_name_zh

WHO country name, formal, Chinese. Only available for official member states.

un_name_en

UN country name, English.

un_name_ru

UN country name, Russian.

un_name_fr

UN country name, French.

un_name_es

UN country name, Spanish.

un_name_ar

UN country name, Arabic.

un_name_zh

UN country name, Chinese.

alt_name_en

Alternative (non-official) country name, English.

alt_name_2_en

Another alternative (non-official) country name, English.

alt_name_3_en

Another alternative (non-official) country name, English.

alt_name_4_en

Another alternative (non-official) country name, English.

alt_name_5_en

Another alternative (non-official) country name, English.

former_name_en

Former country (non-official) name, English.

former_name_2_en

Another former (non-official) country name, English.

who_region

WHO region code.

un_region

UN region code.

un_subregion

UN sub-region code.

un_intermediate_region

UN intermediate region code.

un_region_name_en

UN region name, English.

un_subregion_name_en

UN subregion name, English.

un_intermediate_region_name_en

UN intermediate region name, English.

un_region_name_ru

UN region name, Russian.

un_subregion_name_ru

UN subregion name, Russian.

un_intermediate_region_name_ru

UN intermediate region name, Russian.

un_region_name_fr

UN region name, French.

un_subregion_name_fr

UN subregion name, French.

un_intermediate_region_name_fr

UN intermediate region name, French.

un_region_name_es

UN region name, Spanish.

un_subregion_name_es

UN subregion name, Spanish.

un_intermediate_region_name_es

UN intermediate region name, Spanish.

un_region_name_ar

UN region name, Arabic.

un_subregion_name_ar

UN subregion name, Arabic.

un_intermediate_region_name_ar

UN intermediate region name, Arabic.

un_region_name_zh

UN region name, Chinese.

un_subregion_name_zh

UN subregion name, Chinese.

un_intermediate_region_name_zh

UN intermediate region name, Chinese.

un_desa_region

UN DESA region code.

un_desa_region_name_en

UN DESA region name, English.

un_desa_subregion

UN DESA subregion code.

un_desa_subregion_name_en

UN DESA subregion name, English.

sdg_region

SDG region code.

sdg_region_name_en

SDG region name, English.

sdg_subregion

SDG subregion code.

sdg_subregion_name_en

SDG subregion name, English.

gbd_region

IHME GBD region code.

gbd_region_name_en

IHME GBD region name, English.

gbd_subregion

IHME GBD subregion code.

gbd_subregion_name_en

IHME GBD subregion name, English.

wb_region

WB geographical region. English name as no code provided by WB.

wb_region_name_en

WB geographical region name, English.

wb_ig_1987

World Bank Income Group code, 1987 (FY 1989) classification.

wb_ig_1988

World Bank Income Group code, 1988 (FY 1990) classification.

wb_ig_1989

World Bank Income Group code, 1989 (FY 1991) classification.

wb_ig_1990

World Bank Income Group code, 1990 (FY 1992) classification.

wb_ig_1991

World Bank Income Group code, 1991 (FY 1993) classification.

wb_ig_1992

World Bank Income Group code, 1992 (FY 1994) classification.

wb_ig_1993

World Bank Income Group code, 1993 (FY 1995) classification.

wb_ig_1994

World Bank Income Group code, 1994 (FY 1996) classification.

wb_ig_1995

World Bank Income Group code, 1995 (FY 1997) classification.

wb_ig_1996

World Bank Income Group code, 1996 (FY 1998) classification.

wb_ig_1997

World Bank Income Group code, 1997 (FY 1999) classification.

wb_ig_1998

World Bank Income Group code, 1998 (FY 2000) classification.

wb_ig_1999

World Bank Income Group code, 1999 (FY 2001) classification.

wb_ig_2000

World Bank Income Group code, 2000 (FY 2002) classification.

wb_ig_2001

World Bank Income Group code, 2001 (FY 2003) classification.

wb_ig_2002

World Bank Income Group code, 2002 (FY 2004) classification.

wb_ig_2003

World Bank Income Group code, 2003 (FY 2005) classification.

wb_ig_2004

World Bank Income Group code, 2004 (FY 2006) classification.

wb_ig_2005

World Bank Income Group code, 2005 (FY 2007) classification.

wb_ig_2006

World Bank Income Group code, 2006 (FY 2008) classification.

wb_ig_2007

World Bank Income Group code, 2007 (FY 2009) classification.

wb_ig_2008

World Bank Income Group code, 2008 (FY 2010) classification.

wb_ig_2009

World Bank Income Group code, 2009 (FY 2011) classification.

wb_ig_2010

World Bank Income Group code, 2010 (FY 2012) classification.

wb_ig_2011

World Bank Income Group code, 2011 (FY 2013) classification.

wb_ig_2012

World Bank Income Group code, 2012 (FY 2014) classification.

wb_ig_2013

World Bank Income Group code, 2013 (FY 2015) classification.

wb_ig_2014

World Bank Income Group code, 2014 (FY 2016) classification.

wb_ig_2015

World Bank Income Group code, 2015 (FY 2017) classification.

wb_ig_2016

World Bank Income Group code, 2016 (FY 2018) classification.

wb_ig_2017

World Bank Income Group code, 2017 (FY 2019) classification.

wb_ig_2018

World Bank Income Group code, 2018 (FY 2020) classification.

wb_ig_2019

World Bank Income Group code, 2019 (FY 2021) classification.

wb_ig_2020

World Bank Income Group code, 2020 (FY 2022) classification.

Source

World Health Organization country reference table, accessed September 23 2020.

United Nations Statistics Division, Country or Area Codes for Statistical Use tables; accessed August 15 2020.

World Bank income groups, fiscal years 1989 - 2021, and regions, accessed August 15 2020.

Organisation for Economic Co-operation and Development membership list, accessed 1 February 2021.

Institute for Health Metrics and Evaluation Global Burden of Disease 2019 Socio-Demographic Index data, accessed 1 February 2021.

United Nations Committee for Development Policy, list of least developed countries, accessed 16 February 2021.

United Nations Department of Economic and Social Affairs, Population Division (2020). International Migrant Stock 2020.

Global Burden of Disease 2019 country and regional classifications, Institute for Health Metrics and Evaluation.